requirements.txt
flask
flask-cors
maxminddb-geolite2